merge the latest version
ihealth was maintained by intbot but unfortunately, no staff maintain it
now. in order to maintain the software, we need to merge the intbot
version with the github version. only in this way can we give ihealth a
compelete version control.

add unit test using cmocka, but is not full covered.

add serveral js games source file, but not embed in ihealth.

make changes in database

make changes in ui part